srcs/api-gateway/requirements.txt
flask
requests
pika
python-dotenv
psycopg2-binary